start ROS API via pass thru encoder layer
the goal is to expose input/output data of encoders/SP/TP/classifier
to ROS.

We'll introduce a fake neutral layer - a passthru encoder which keeps all data intact,
but will stream them to ROS as well. These layer can be placed at any place in the region structure.

For ROS nodes, the Pubisher/Listener functionality is implemented, not service