Spatial pooler bug fixes and improvements
Fixed bug in initialization of minActiveDutyCycles. This bug was causing boosting to occur very aggressively in the beginning. This in turn causes the SP to learn bad SDR’s.

Added separate spatial_pooler_compute_test.py for doing end to end tests of the compute function. We primarily had unit tests of the individual methods before with no tests of compute. This test is relatively basic but will catch some failures that the other tests do not.

Started implementing a test of SP boosting. Currently only Phase 1 is implemented.

Added printParameters() to the python spatial pooler to match the CPP ones. This is helpful for debugging.

The new spatial poolers will now print out the main construction parameters if spVerbosity >= 1

Improved some of the comments.