Boost test complete (hopefully)
Fixed bug in how the tie breakers are used in do inhibition. In the implementation a random tie breaker is generated on every iteration. This leads to different SDRs for the same input, even if no learning or other changes have occurred. We do need a small random tie breaker, but the random offsets can be set during initialization.

Implemented phase 4 of boost test. This test is now complete.   Made test a bit more robust to randomness by increasing the number of columns. Turned learning off to isolate boosting related effects.