removed strstream from SharedPythonOStream
This commit includes what seems like an unrelated change in svm.cpp, but it seems that
the previous SharedPythonOStream implementation was making up for a misreported persistent_size
in the svm_parameter object.

I believe the final two string recastings of close in the SharedPythonOStream class to be the
place where the string length got adjusted, but I'm not sure exactly how that worked.

Furthermore, the SharedPythonOStream object has undergone substantial changes because it
was previously internally sharing a character buffer between a strstream and a PyString object.
Using stringstream, this is not possible, so any code relating to it is removed.